chore: combine latest column defs for market list changes to market code col
